+++ src/clikraken/api/private/list_open_orders.py
@@ -14,6 +14,7 @@ from decimal import Decimal
 from clikraken.api.api_utils import parse_order_res, query_api
 from clikraken.clikraken_utils import asset_pair_short
 from clikraken.clikraken_utils import _tabulate as tabulate
+from clikraken.clikraken_utils import csv
 
 
 def list_open_orders(args):
@@ -23,11 +24,13 @@ def list_open_orders(args):
     api_params = {
         # TODO
     }
-
-    res = query_api('private', 'OpenOrders', api_params, args)
-
-    # extract list of orders from API results
-    res_ol = res['open']
+    if args.txid:
+        api_params.update({'txid': args.txid})
+        res_ol = query_api('private', 'QueryOrders', api_params, args)
+    else:
+        res = query_api('private', 'OpenOrders', api_params, args)
+        # extract list of orders from API results
+        res_ol = res['open']
 
     # the parsing is done in an helper function
     ol = parse_order_res(res_ol, ['open'])
@@ -47,4 +50,7 @@ def list_open_orders(args):
     if not ol_all:
         return
 
-    print(tabulate(ol_all, headers="keys"))
+    if args.csv:
+        print(csv(ol_all, headers="keys"))
+    else:
+        print(tabulate(ol_all, headers="keys"))