+++ src/clikraken/api/public/depth.py
@@ -15,6 +15,7 @@ from decimal import Decimal
 from clikraken.api.api_utils import query_api
 from clikraken.clikraken_utils import asset_pair_short, humanize_timestamp
 from clikraken.clikraken_utils import _tabulate as tabulate
+from clikraken.clikraken_utils import csv
 
 
 def depth(args):
@@ -55,7 +56,23 @@ def depth(args):
         # sort by price descending
         depth_dict[dtype] = reversed(sorted(depth_dict[dtype], key=lambda dentry: Decimal(dentry[price_label])))
 
-    asks_table = tabulate(depth_dict['asks'], headers="keys")
-    bids_table = tabulate(depth_dict['bids'], headers="keys")
-
-    print("{}\n\n{}".format(asks_table, bids_table))
+    if args.csv:
+        output = []
+        for dtype in depth_dict.keys():
+            for o in depth_dict[dtype]:
+                it = OrderedDict()
+                it['dtype'] = dtype
+                for k, v in o.items():
+                    if len(k.split(' ')) > 1:
+                        # key has a space, this is the "price_label" column -> "XABCZDEF Ask"
+                        it['pair'] = k.split(' ')[0]  # keep only "XABCZDEF"
+                        it['price'] = v
+                    else:
+                        # the other columns don't contain a space
+                        it[k] = v
+                output += [it]
+        print(csv(output, headers="keys"))
+    else:
+        asks_table = tabulate(depth_dict['asks'], headers="keys")
+        bids_table = tabulate(depth_dict['bids'], headers="keys")
+        print("{}\n\n{}".format(asks_table, bids_table))