+++ src/clikraken/global_vars.py
@@ -28,7 +28,7 @@ DEFAULT_SETTINGS_INI = """[clikraken]
 # and the environment variable CLIKRAKEN_DEFAULT_PAIR is not set
 currency_pair = XETHZEUR
 
-# default currency pair when no option '-a' or '--asset' is given
+# default asset when no option '-a' or '--asset' is given
 # and the environment variable CLIKRAKEN_DEFAULT_ASSET is not set
 asset = ETH
 
@@ -54,3 +54,4 @@ TZ = None
 TRADING_AGREEMENT = None
 CRON = None
 API_KEY_LOADED = None
+CSV_SEPARATOR = None